pyproject.toml
[project]
name = "ez-weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "openai>=1.3.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "pydantic>=2.10.6",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
